visual_relation/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
numpy>=1.19.5
pandas>=1.4.4
torch==1.13.1+cu117
torchvision==0.14.1+cu117
Pillow>=10.3.0
tensorboardX>=2.6.2.2
open-clip-torch>=2.24.0
gdown>=5.1.0
svn
